MAME 0.102u2

+++ src/sndhrdw/redbaron.c
@@ -21,7 +21,7 @@
 #include "sound/pokey.h"
 
 /* used in drivers/redbaron.c to select joystick pot */
-extern int rb_input_select;
+extern UINT8 rb_input_select;
 
 /* Statics */
 static INT16 *vol_lookup = NULL;
